requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
click==8.2.1
-e git+https://github.com/ghombal/data_driven.git@f3720d59e9d6fe575fa1ab42c9625d98c8a269eb#egg=data_driven
eventkit==1.0.3
frozenlist==1.7.0
ib-insync==0.9.86
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-mock==3.14.1
python-dotenv==1.1.1
PyYAML==6.0.2
SQLAlchemy==2.0.41
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
yarl==1.20.1